requirements.txt
Django>=5.0
requests>=2.31.0
openai>=1.0.0
celery>=5.3.0
kombu>=5.3.0
python-decouple>=3.8
django-extensions>=3.2.0
Pillow>=10.0.0
# Production WSGI server
gunicorn>=21.2.0
# Database
psycopg2-binary>=2.9.0
dj-database-url>=2.1.0
# Celery brokers
redis>=4.5.0
# Celery monitoring
flower>=2.0.0
# Django Celery integration
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
google-genai>=0.3.0
# Cache
django-redis>=5.4.0
# PDF Reports
reportlab>=4.0.0